requirements.txt
Flask
Flask-Cors
pillow
pydub
numpy
torch
torchvision
urllib3==1.26.6